Revised the error part and fetched datas in sql in main.py def dash to be specific also fix an error on the warning in the bottom part of the script
