--- INVENTORYMSNIKIZAYA/IMS JAKV/haha.html
@@ -1,239 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Add Product</title>
-    <style>
-        body {
-            font-family: Arial, sans-serif;
-            display: flex;
-            justify-content: center;
-            align-items: center;
-            height: 100vh;
-            margin: 0;
-            background-color: #85FFBD;
-            background-image: linear-gradient(45deg, #85FFBD 0%, #FFFB7D 100%);
-
-        }
-
-        .container {
-            background-color: #FBAB7E;
-            background-image: linear-gradient(62deg, #FBAB7E 0%, #F7CE68 100%);
-            padding: 20px;
-            border-radius: 10px;
-            box-shadow: 0 8px 16px rgba(0, 0, 0, 0.2);
-            max-width: 400px;
-            width: 100%;
-        }
-
-        h1 {
-            text-align: center;
-            margin-bottom: 20px;
-            font-weight: bold;
-        }
-
-        .form-group {
-            margin-bottom: 15px;
-            font-weight: bold;
-            font-family: Tahoma, Geneva, Verdana, sans-serif;
-        }
-
-        .form-group label {
-            display: block;
-            margin-bottom: 5px;
-        }
-
-        .form-group input {
-            width: 95%;
-            padding: 10px;
-            border: 1px solid #ddd;
-            border-radius: 5px;
-        }
-
-        .form-group button {
-            width: 100%;
-            padding: 10px;
-            border: none;
-            border-radius: 5px;
-            background-color: #70b11b;
-            color: #000000;
-            font-size: 16px;
-            cursor: pointer;
-            font-weight: bold;
-        }
-
-        .form-group button:hover {
-            background-color: #504717;
-        }
-    </style>
-</head>
-
-<body>
-    <div class="container">
-        <h1>Add Product</h1>
-        <form id="addProductForm">
-            <div class="form-group">
-                <label for="productId">Product ID</label>
-                <input type="text" id="productId" name="productId" placeholder="Enter product ID" required>
-            </div>
-            <div class="form-group">
-                <label for="productName">Product Name</label>
-                <input type="text" id="productName" name="productName" placeholder="Enter product name" required>
-            </div>
-            <div class="form-group">
-                <label for="productPrice">Product Price</label>
-                <input type="number" id="productPrice" name="productPrice" placeholder="Enter product price" step="0.01"
-                    required>
-            </div>
-            <div class="form-group">
-                <label for="productDescription">Product Stock</label>
-                <input type="number" id="productDescription" name="productDescription"
-                    placeholder="Enter product stock" required>
-            </div>
-            <div class="form-group">
-                <button type="submit">Add Product</button>
-            </div>
-        </form>
-    </div>
-</body>
-
-</html>
-
-
-
-
-
-
-
-
-
-
-
-
-.main_content .button{
-    color: #382a2a;
-    padding: 10px 20px;
-    border: none;
-    cursor: pointer;
-    transition: background-color 0.3s ease;
-}
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="utf-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0, maximum-scale=1.0, user-scalable=no">
-    <title>INVENTORY</title>
-    <link rel="stylesheet" href="static/inv.css" >
-
-  </head>
-  <body>
-
-
-
-       
-
-  <div class="wrapper">
-    <div class="sidebar">
-      <h3>INVENTORY MANAGEMENT</h3>
-      <ul>
-
-
-
-
-
-        <li><a href="{{url_for('dashboard')}}">  DASHBOARD  </a></li>
-
-        <li><a href="{{url_for('inventory')}}">  INVENTORY  </a></li>
-
-        <li><a href="{{url_for('orders')}}">  ORDERS  </a></li>
-
-        <li><a href="{{url_for('aboutus')}}">  ABOUT US  </a></li>
-
-        
-
-
-
-
-      </ul>
-      
-
-    </div>
-
-
-    <div class="main_content">
-      <div class="header">INVENTORY</div>
-      <div class="info">
-        <div>you can see here whats in inventory</div>
-
-        {% for i in inv %}
-
-
-
-        <p>{{i[0]}} {{i[1]}}  {{i[2]}}  {{i[3]}}</p>
-
-        {% endfor %}
-
-
-
-
-      </div>
-    </div>
-  
-  </div>
-  
-
-
-
-
-
-
-
-
-  </body>
-</html>
