--- INVENTORYMSNIKIZAYA/IMS JAKV/static/update.css
@@ -1,59 +1,0 @@
-*{
-    margin: 0;
-    padding: 0;
-    box-sizing: border-box;
-    list-style: none;
-    text-decoration: none;
-}
-
-body {
-    background-color: #FADCD5;
-}
-
-h1 {
-    font-family: "Cormorant Garamond", serif;
-    font-weight: 300;
-    font-style: normal;
-    text-align: center;
-    padding-bottom: 30px;
-    color: #000000;
-    font-family: "Kanit", sans-serif;
-    font-weight: 400;
-    font-style: normal;
-    margin-top: 60px;
-}
-
-form {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-}
-
-input[type="text"] {
-    width: 400px;
-    padding: 20px;
-    margin-bottom: 10px;
-    border: none;
-    border-radius: 5px;
-    outline: 1px solid #765D67; 
-}
-
-button[type="submit"] {
-    background-color: #4B2138; 
-    color: #fff;
-    padding: 10px 20px;
-    border: none;
-    border-radius: 5px;
-    cursor: pointer;
-}
-
-button[type="submit"]:hover {
-    background-color: #765D67; 
-}
-
-
-
-
-
- 
-
