--- INVENTORYMSNIKIZAYA/IMS JAKV/templates/dash.html
@@ -1,107 +1,0 @@
-<!DOCTYPE html> 
-<html lang="en"> 
-  <head> 
-    <meta charset="utf-8"> 
-    <meta name="viewport" content="width=device-width, initial-scale=1.0, maximum-scale=1.0, user-scalable=no"> 
-    <title>INVENTORY</title> 
-    <link rel="stylesheet" href="static/inv.css" > 
- 
-    <link rel="preconnect" href="https://fonts.googleapis.com"> 
-    <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin> 
-    <link href="https://fonts.googleapis.com/css2?family=Kanit:ital,wght@0,100;0,200;0,300;0,400;0,500;0,600;0,700;0,800;0,900;1,100;1,200;1,300;1,400;1,500;1,600;1,700;1,800;1,900&display=swap" rel="stylesheet"> 
-    <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.1.1/css/all.min.css" /> 
-     
- 
- 
- 
-  </head> 
-  <body> 
-    <div class="hamburger-menu"> 
-        <i class="fas fa-bars"></i> 
-    </div> 
-    <div class="wrapper"> 
-      <div class="sidebar"> 
-        <h3>INVENTORY MANAGEMENT</h3> 
-        <ul> 
-          <li><a href="{{url_for('dashboard')}}"> <i class="fa-solid fa-chart-pie"></i> DASHBOARD</a></li>
-          <li><a href="{{url_for('inventory')}}"> <i class="fa-solid fa-boxes-stacked"></i>  INVENTORY  </a></li> 
-          <li><a href="{{url_for('addprd')}}"> <i class="fa-solid fa-square-plus"></i> ADD PRODUCT</a></li>
-          <li><a href="{{url_for('transaction')}}"><i class="fa-solid fa-clipboard-list"></i>  TRANSACTIONS </a></li>
-          <li><a href="{{url_for('aboutus')}}"> <i class="fa-regular fa-address-card"></i> ABOUT US  </a></li> 
-          <li><a href="{{url_for('logout')}}"> <i class="fa-solid fa-right-from-bracket"></i> LOGOUT  </a></li> 
-        </ul> 
-      </div> 
-      <div class="main_content"> 
-        <div class="header">DASHBOARD</div> 
-        <div class="info"> 
-
-
-            <script type="text/javascript" src="https://www.gstatic.com/charts/loader.js"></script>
-            <script type="text/javascript">
-                google.charts.load('current', {'packages':['corechart']});
-                google.charts.setOnLoadCallback(drawChart);
-
-      function drawChart() {
-
-        var data = google.visualization.arrayToDataTable([
-          ['Product', 'Quantity'],
-          ['Drinks', 40],
-          ['Snacks', 170],
-          ['Daily Products', 100],
-          ['Condiments and Sauces', 100],
-          ['Household Items', 100],
-          ['Personal Care Products', 120],
-        ]);
-
-        
-
-
-
-        var options = {
-          title: 'PRODUCT OVERALL STOCKS',
-          backgroundColor: '#dfeae9',
-                pieSliceText: 'label',
-                legend: {
-                    position: 'bottom',
-                    alignment: 'center',
-                    textStyle: {
-                        color: 'black',
-                    }
-                }
-        };
-
-        var chart = new google.visualization.PieChart(document.getElementById('piechart'));
-
-         // Add warning sign for low stock
-         for (var i = 0; i < data.getNumberOfRows(); i++) {
-          if (data.getValue(i, 1) < 50) { // Check if quantity is less than 50
-            var warningSign = '<span style="color: red; animation: blink 1s infinite;">⚠️</span>'; // Warning sign symbol
-            data.setValue(i, 0, data.getValue(i, 0) + warningSign); // Add warning sign to product label
-          }
-         }
-      
-
-  
-
-        chart.draw(data, options);
-      }
-    </script>
-    <div id="piechart" style="width: 900px; height: 500px;"></div>
-
-
-
-
- 
-    <script> 
-        document.querySelector('.hamburger-menu').addEventListener('click', function () { 
-          document.querySelector('.sidebar').classList.toggle('active'); 
-        }); 
-      </script> 
-
-        </div>
-    </div>
-
-
-
-  </body> 
-</html>
