--- INVENTORYMSNIKIZAYA/IMS JAKV/templates/login.html
@@ -1,62 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="utf-8">
-    <title>LOGIN</title>
-    <link rel="stylesheet" href="static/login.css" >
-    <link rel="preconnect" href="https://fonts.googleapis.com">
-    <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
-    <link href="https://fonts.googleapis.com/css2?family=Oswald:wght@200..700&display=swap" rel="stylesheet">
-  </head>
-  <body>
-    
-    
-
-    <div class="header">
-      <h1>WELCOME TO INVENTORY MANAGEMENT SYSTEM.</h1>
-    </div>
-
-
-    
-
-
-
-
-
-    <form action="/login_process" method="post">
-      {% with messages = get_flashed_messages() %}
-        {% if messages %}
-          <ul class="flash-messages">
-        
-      
-            {% for message in messages %}
-              <li>{{ message }}</li>
-            {% endfor %}
-          </ul>
-        {% endif %}
-      {% endwith %}
-      <h3>PLEASE LOG IN.</h3>
-        <input type="text" placeholder="username" name="userfield" required>
-        <br>
-        <input type="password" placeholder="password" name="passfield" required>
-        <br>
-        <p>Don't have an account? <a href="{{ url_for('signup') }}">Sign up here</a></p>
-        <button> login</button>
-        
-        
-    </form>
-
-
-
-
-    
-
-
-
-
-
-
-
-
-  </body>
-</html>
