--- INVENTORYMSNIKIZAYA/IMS JAKV/templates/search_res.html
@@ -1,78 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="utf-8">
-    <link rel="stylesheet" href="static/search.css">
-    
-    <link rel="preconnect" href="https://fonts.googleapis.com">
-    <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
-    <link href="https://fonts.googleapis.com/css2?family=Kanit:ital,wght@0,100;0,200;0,300;0,400;0,500;0,600;0,700;0,800;0,900;1,100;1,200;1,300;1,400;1,500;1,600;1,700;1,800;1,900&display=swap" rel="stylesheet">
-
-    <title>SEARCH</title>
-  </head>
-  <body>
-    <div class="info">
-      <h2>Search Results</h2>
-      <table class="search-results-table">
-        <thead>
-          <tr>
-            <th>Product ID</th>
-            <th>Product Name</th>
-            <th>Product Quantity</th>
-            <th>Product Price</th>
-            <th>Operation</th>
-          </tr>
-        </thead>
-        <tbody>
-          {% for product in search_results %}
-            <tr>
-              <td>{{ product[0] }}</td>
-              <td>{{ product[1] }}</td>
-              <td>{{ product[2] }}</td>
-              <td>{{ product[3] }}</td>
-              <td>
-                <div class="action-buttons">
-                  <form action="/update" method="post">
-                    <input type="hidden" name="productID" value="{{ product[0] }}">
-                    <input type="hidden" name="product_name" value="{{ product[1] }}">
-                    <input type="hidden" name="Qty" value="{{ product[2] }}">
-                    <input type="hidden" name="price" value="{{ product[3] }}">
-                    <button type="submit" name="action" value="updater">update product</button>
-                  </form>
-                  <form action="/update" method="post">
-                    <input type="hidden" name="productID" value="{{ product[0] }}">
-                    <button type="submit" name="action" value="deleter">remove product</button>
-                  </form>
-                </div>
-              </td>
-            </tr>
-          {% endfor %}
-        </tbody>
-      </table>
-    </div>
-   
-
-  
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-    
-  </body>
-</html>
