--- INVENTORYMSNIKIZAYA/IMS JAKV/templates/update.html
@@ -1,25 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="utf-8">
-    <link rel="preconnect" href="https://fonts.googleapis.com">
-    <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
-    <link href="https://fonts.googleapis.com/css2?family=Kanit:ital,wght@0,100;0,200;0,300;0,400;0,500;0,600;0,700;0,800;0,900;1,100;1,200;1,300;1,400;1,500;1,600;1,700;1,800;1,900&display=swap" rel="stylesheet">
-    <title>UPDATE</title>
-    <link rel="stylesheet" href="static/update.css" >
-  </head>
-  <body>
-    <h1>YOU ARE UPDATING PRODUCT {{product_name}}</h1>
-    <form action="/update_process" method="POST">
-        <input type="text" name="productID" value="{{productID}}">
-        <br>
-        <input type="text" name="new_product_name" value="{{product_name}}">
-        <br>
-        <input type="text" name="new_productQuantity" value="{{productQuantity}}">
-        <br>
-        <input type="text" name="new_product_price" value="{{product_price}}">
-        <br>
-        <button type="submit"> UPDATE PRODUCT </button>
-    </form>
-  </body>
-</html>
