+++ static/aboutus.css
@@ -1,0 +1,256 @@
+* {
+    margin: 0;
+    padding: 0;
+    box-sizing: border-box;
+    list-style: none;
+    text-decoration: none;
+}
+
+
+
+
+body {
+    background-color: #dfeae9;
+}
+
+h3 {
+    text-align: center;
+    padding-bottom: 30px;
+    color: #ffffff;
+    font-family: "Kanit", sans-serif;
+    font-weight: 400;
+    font-style: normal;
+}
+
+.wrapper {
+    display: flex;
+    position: relative;
+}
+
+.wrapper .sidebar {
+    position: fixed;
+    width: 205px;
+    height: 100%;
+    background-color: #496246;
+    padding: 30px 0;
+}
+
+.wrapper .sidebar h1 {
+    text-transform: uppercase;
+    text-align: center;
+    margin-bottom: 30px;
+    font-style: normal;
+    font-family: "Kanit", sans-serif;
+    font-weight: 400;
+    font-style: normal;
+    font-size: 200%;
+    
+}
+
+.wrapper .sidebar ul li {
+    padding: 15px;
+    border-bottom: 1px solid  #496246;
+    border-top: 1px solid  #496246;
+}
+
+
+
+.wrapper .sidebar ul li a {
+    color: #fff;
+    display: block;
+    font-family: "Kanit", sans-serif;
+    font-weight: 400;
+    font-style: normal;
+}
+
+
+
+.wrapper .sidebar ul li:hover {
+    background: #97a78c;
+}
+
+
+
+
+.wrapper .sidebar ul li:hover a {
+    color: #000000;
+}
+
+
+
+.wrapper .main_content {
+    width: 100%;
+    margin-left: 205px; 
+}
+
+
+
+.wrapper .main_content .header { 
+  padding: 30px; 
+  background-color: #97b6b1; 
+  color: #000000; 
+  border-bottom: 5px solid #496246; 
+  text-align: center; 
+  font-size: 200%; 
+  font-weight: bolder; 
+  font-family: "Kanit", sans-serif; 
+  font-weight: 400; 
+  font-style: normal; 
+} 
+
+.wrapper .main_content .info {
+    margin: 20px;
+    line-height: 20px;
+}
+
+.wrapper .main_content .info div {
+    margin-bottom: 20px;
+}
+
+
+
+
+
+.main_content .info h2{
+  margin: 20px;
+  font-family: "Kanit", sans-serif;
+  font-weight: 400;
+  font-style: normal;
+}
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+.team {
+    display: flex;
+    flex-wrap: wrap;
+    justify-content: center;
+  }
+  
+  .leader {
+    font-family: "Kanit", sans-serif;
+    font-weight: 400;
+    font-style: normal;
+    background-color: #6c8969; 
+    padding: 20px;
+    margin: 20px;
+    width: 200px;
+    border-radius: 10px;
+    box-shadow: 0 0 10px rgba(0, 0, 0, 0.2);
+  }
+  
+  .leader-box {
+    padding: 20px;
+  }
+  
+  .members {
+    display: flex;
+    flex-wrap: wrap;
+    justify-content: center;
+  }
+  
+  .member-box {
+    background-color: #6c8969; 
+    padding: 20px;
+    margin: 20px;
+    width: 200px;
+    border-radius: 10px;
+    box-shadow: 0 0 10px rgba(0, 0, 0, 0.2);
+  }
+  
+  .member-box h3 {
+    margin-top: 0;
+    font-family: "Kanit", sans-serif;
+    font-weight: 400;
+    font-style: normal;
+  }
+  
+  .member-box p {
+    margin-bottom: 10px;
+    font-family: "Kanit", sans-serif;
+    font-weight: 400;
+    font-style: normal;
+  }
+
+
+
+
+  
+ 
+ 
+ 
+  .wrapper .main_content .info ul { 
+    text-align: center; 
+  } 
+   
+  .wrapper .main_content .info ul li { 
+    color: green; 
+    font-size: large; 
+  } 
+  
+  .hamburger-menu { 
+  display: none;  
+  position: fixed; 
+  top: 5px; 
+  left: 15px; 
+  background-color: #496246; 
+  padding: 10px; 
+  border-radius: 5px; 
+  cursor: pointer; 
+  z-index: 1000; 
+  } 
+  
+  .hamburger-menu i { 
+  color:#fff; 
+  font-size: 24px; 
+  } 
+  
+  @media (max-width: 768px) { 
+  .hamburger-menu { 
+    display: block; 
+  } 
+  
+  .sidebar { 
+    position: fixed; 
+    left: -205px;  
+    transition: left 0.3s; 
+  } 
+  
+  .sidebar.active { 
+    left: 0;  
+  } 
+  
+  .wrapper .main_content { 
+    margin-left: 0; 
+    padding-top: 50px; 
+  } 
+  } 
+  @media (max-width: 768px) { 
+  .search-bar { 
+    position: static; 
+    margin: 20px auto; 
+    text-align: center;  
+    width: 100%; 
+  } 
+  .wrapper .main_content { 
+    padding-top: 70px;  
+  } 
+  } 
+  @media (max-width: 768px) { 
+  .wrapper .main_content { 
+    padding-top: 0px; 
+  }
+}
