+++ static/login.css
@@ -1,0 +1,116 @@
+* {
+  margin: 0;
+  padding: 0;
+  box-sizing: border-box;
+  list-style: none;
+  text-decoration: none;
+}
+
+body {
+  background-image: url('green.jpg');
+  background-size: cover;
+  background-position: center;
+  height: 100vh;
+  margin: 0;
+}
+
+.header {
+  text-align: center;
+  margin-top: 70px;
+  margin-bottom: 50px;
+  font-family: "Oswald", sans-serif;
+  font-optical-sizing: auto;
+  font-weight: 500;
+  font-style: normal;
+  font-size: x-large;
+  color: rgb(114, 175, 117);
+}
+
+form {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  padding: 20px;
+  border-radius: 10px;
+  width: 500px;
+  height: 300px;
+  position: absolute;
+  top: 50%;
+  left: 50%;
+  transform: translate(-50%, -50%);
+  margin: 0;
+  background: rgba(55,130,205,0.1);
+  -webkit-backdrop-filter: blur(6px);
+  backdrop-filter: blur(6px);
+  border: 1px solid rgba(55,130,205,0.05);
+}
+
+form input {
+  padding: 10px;
+  width: 100%;
+  height: 100px;
+  border: none;
+  border-radius: 5px;
+  outline: 1px solid #4B2138;
+  margin-top: 30px;
+}
+
+form button {
+  background-color: #4B2138;
+  color: #fff;
+  padding: 10px 20px;
+  border: none;
+  border-radius: 5px;
+  cursor: pointer;
+  transition: background-color 0.3s ease;
+}
+
+form button:hover {
+  background-color: #765D67;
+}
+
+.flash-messages {
+  list-style: none;
+  padding: 10px;
+  background-color: #C6F4D6; 
+  border: 1px solid #3E8E41; 
+  border-radius: 5px;
+  color: #3E8E41; 
+  text-align: center;
+  margin-bottom: 20px;
+}
+
+.flash-messages li {
+  margin-bottom: 10px;
+}
+
+@media only screen and (min-width: 768px) and (max-width: 1024px) {
+  form {
+    width: 80%;
+    height: 250px;
+  }
+  form input {
+    height: 80px;
+  }
+}
+
+
+@media only screen and (max-width: 767px) {
+  form {
+    width: 90%;
+    height: 200px;
+    box-shadow: none; 
+    padding-bottom: 20px; 
+  }
+  form input {
+    height: 60px;
+  }
+  .header {
+    margin-top: 40px;
+    margin-bottom: 30px;
+    font-size: large; 
+  }
+  form button {
+    margin-top: 10px; 
+  }
+}
