+++ static/search.css
@@ -1,0 +1,85 @@
+.search-results-table {
+    border-collapse: collapse;
+    width: 100%;
+  }
+  
+  .search-results-table th,
+  .search-results-table td {
+    border: 1px solid #000000;
+    padding: 8px;
+    text-align: center;
+    font-family: "Kanit", sans-serif;
+    font-weight: 400;
+    font-style: normal;
+  }
+  
+  .search-results-table th {
+    background-color: #2D222F; 
+    color: #fff; 
+    font-family: "Kanit", sans-serif;
+    font-weight: 400;
+    font-style: normal;
+  }
+  
+  .search-results-table th:nth-child(1),
+  .search-results-table th:nth-child(2),
+  .search-results-table th:nth-child(3),
+  .search-results-table th:nth-child(4) {
+    background-color: #2D222F; 
+    color: #fff; 
+    font-family: "Kanit", sans-serif;
+    font-weight: 400;
+    font-style: normal;
+  }
+
+h2{
+    text-align: center;
+
+}
+
+
+
+  body{
+    background-color:#FADCD5 ;
+    margin: 0;
+    padding: 0;
+    font-family: "Kanit", sans-serif;
+    font-weight: 400;
+    font-style: normal;
+  }
+
+
+
+
+
+  button[name="action"][value="updater"] {
+    background-color: #007200; 
+    color: #ffffff;
+    border: none;
+    padding: 5px 10px;
+    font-size: 16px;
+    cursor: pointer;
+    margin-bottom: 10px;
+  }
+  
+  button[name="action"][value="updater"]:hover {
+    background-color: #008000; 
+  }
+  
+ 
+  button[name="action"][value="deleter"] {
+    background-color: #d00000; 
+    color: #ffffff;
+    border: none;
+    padding: 5px 10px;
+    font-size: 16px;
+    cursor: pointer;
+  }
+  
+  button[name="action"][value="deleter"]:hover {
+    background-color: #9d0208;
+  }
+
+
+
+
