+++ static/trans.css
@@ -1,0 +1,267 @@
+* { 
+    margin: 0; 
+    padding: 0; 
+    box-sizing: border-box; 
+    list-style: none; 
+    text-decoration: none; 
+  } 
+   
+  body { 
+    background-color: #dfeae9; 
+  } 
+  
+  h3 { 
+    font-family: "Cormorant Garamond", serif; 
+    font-weight: 300; 
+    font-style: normal; 
+    text-align: center; 
+    padding-bottom: 30px; 
+    color: #ffffff; 
+    font-family: "Kanit", sans-serif; 
+    font-weight: 400; 
+    font-style: normal; 
+  } 
+  .wrapper { 
+    display: flex; 
+    position: relative; 
+  } 
+  .wrapper .sidebar { 
+    position: fixed; 
+    width: 205px; 
+    height: 100%; 
+    background-color: #496246; 
+    padding: 30px 0; 
+  } 
+  .wrapper .sidebar h1 { 
+    text-transform: uppercase; 
+    text-align: center; 
+    margin-bottom: 30px; 
+    font-style: normal; 
+    font-family: "Kanit", sans-serif; 
+    font-weight: 400; 
+    font-style: normal; 
+    font-size: 200%; 
+     
+  } 
+  .wrapper .sidebar ul li { 
+    padding: 15px; 
+    border-bottom: 1px solid  #496246; 
+    border-top: 1px solid  #496246; 
+  } 
+   
+   
+  .wrapper .sidebar ul li a { 
+    color: #fff; 
+    display: block; 
+    font-family: "Kanit", sans-serif; 
+    font-weight: 400; 
+    font-style: normal; 
+  } 
+   
+   
+  .wrapper .sidebar ul li:hover { 
+    background: #97a78c; 
+  } 
+   
+   
+   
+  .wrapper .sidebar ul li:hover a { 
+    color: #000000; 
+  } 
+  .wrapper .main_content { 
+    width: 100%; 
+    margin-left: 205px;  
+  } 
+   
+   
+  .wrapper .main_content .header { 
+    padding: 30px; 
+    background-color: #97b6b1; 
+    color: #000000; 
+    border-bottom: 5px solid #496246; 
+    text-align: center; 
+    font-size: 200%; 
+    font-weight: bolder; 
+    font-family: "Kanit", sans-serif; 
+    font-weight: 400; 
+    font-style: normal; 
+  } 
+  .wrapper .main_content .info { 
+    margin: 20px; 
+    line-height: 20px; 
+  } 
+  .wrapper .main_content .info div { 
+    margin-bottom: 20px; 
+  } 
+   
+  .main_content .info h2{ 
+    margin: 20px; 
+    font-family: "Kanit", sans-serif; 
+    font-weight: 400; 
+    font-style: normal; 
+  } 
+   
+   
+   
+  table { 
+    border-collapse: collapse; 
+    width: 100%; 
+  } 
+   
+  th, td { 
+    border: 1px solid #496246; 
+    padding: 8px; 
+    text-align: center; 
+    font-family: "Kanit", sans-serif; 
+    font-weight: 400; 
+    font-style: normal; 
+  } 
+   
+   
+  th { 
+    background-color: #496246;  
+    color: #fff;  
+    font-family: "Kanit", sans-serif; 
+    font-weight: 400; 
+    font-style: normal; 
+  } 
+   
+  th label { 
+    font-weight: bold; 
+     
+  } 
+   
+   
+   
+   
+  button[name="action"][value="updater"] { 
+    background-color: #007200;  
+    color: #ffffff; 
+    border: none; 
+    padding: 10px 20px; 
+    font-size: 16px; 
+    cursor: pointer; 
+  } 
+   
+  button[name="action"][value="updater"]:hover { 
+    background-color: #008000;  
+  } 
+   
+   
+  button[name="action"][value="deleter"] { 
+    background-color: #d00000;  
+    color: #ffffff; 
+    border: none; 
+    padding: 10px 20px; 
+    font-size: 16px; 
+    cursor: pointer; 
+  } 
+  button[name="action"][value="deleter"]:hover { 
+    background-color: #9d0208; 
+  } 
+   
+   
+   
+   
+   
+   
+   
+  .search-bar { 
+    position: absolute; 
+    right: 20px; 
+     
+  } 
+   
+   
+   
+  .search-bar input[type="text"] { 
+    padding: 10px; 
+    border: none; 
+    border-radius: 20px; 
+    width: 200px; 
+    background-color: #ffffff; 
+     
+  } 
+   
+   
+   
+  .search-bar button[type="submit"] { 
+    padding: 10px 20px; 
+    border: none; 
+    border-radius: 20px; 
+    background-color: #496246; 
+    color: #ffffff; 
+    cursor: pointer; 
+    width: 90px; 
+  } 
+   
+   
+  .search-bar button[type="submit"]:hover { 
+    background-color: #496246; 
+  } 
+   
+   
+   
+   
+  .wrapper .main_content .info ul { 
+    text-align: center; 
+  } 
+   
+  .wrapper .main_content .info ul li { 
+    color: green; 
+    font-size: large; 
+  } 
+  
+  .hamburger-menu { 
+  display: none;  
+  position: fixed; 
+  top: 5px; 
+  left: 15px; 
+  background-color: #496246; 
+  padding: 10px; 
+  border-radius: 5px; 
+  cursor: pointer; 
+  z-index: 1000; 
+  } 
+  
+  .hamburger-menu i { 
+  color:#fff; 
+  font-size: 24px; 
+  } 
+  
+  @media (max-width: 768px) { 
+  .hamburger-menu { 
+    display: block; 
+  } 
+  
+  .sidebar { 
+    position: fixed; 
+    left: -205px;  
+    transition: left 0.3s; 
+  } 
+  
+  .sidebar.active { 
+    left: 0;  
+  } 
+  
+  .wrapper .main_content { 
+    margin-left: 0; 
+    padding-top: 50px; 
+  } 
+  } 
+  @media (max-width: 768px) { 
+  .search-bar { 
+    position: static; 
+    margin: 20px auto; 
+    text-align: center;  
+    width: 100%; 
+  } 
+  .wrapper .main_content { 
+    padding-top: 70px;  
+  } 
+  } 
+  @media (max-width: 768px) { 
+  .wrapper .main_content { 
+    padding-top: 0px; 
+  }
+  }
