+++ templates/dash.html
@@ -1,0 +1,105 @@
+<!DOCTYPE html> 
+<html lang="en"> 
+  <head> 
+    <meta charset="utf-8"> 
+    <meta name="viewport" content="width=device-width, initial-scale=1.0, maximum-scale=1.0, user-scalable=no"> 
+    <title>INVENTORY</title> 
+    <link rel="stylesheet" href="static/inv.css" > 
+ 
+    <link rel="preconnect" href="https://fonts.googleapis.com"> 
+    <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin> 
+    <link href="https://fonts.googleapis.com/css2?family=Kanit:ital,wght@0,100;0,200;0,300;0,400;0,500;0,600;0,700;0,800;0,900;1,100;1,200;1,300;1,400;1,500;1,600;1,700;1,800;1,900&display=swap" rel="stylesheet"> 
+    <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.1.1/css/all.min.css" /> 
+     
+ 
+ 
+ 
+  </head> 
+  <body> 
+    <div class="hamburger-menu"> 
+        <i class="fas fa-bars"></i> 
+    </div> 
+    <div class="wrapper"> 
+      <div class="sidebar"> 
+        <h3>INVENTORY MANAGEMENT</h3> 
+        <ul> 
+          <li><a href="{{url_for('dashboard')}}"> <i class="fa-solid fa-chart-pie"></i> DASHBOARD</a></li>
+          <li><a href="{{url_for('inventory')}}"> <i class="fa-solid fa-boxes-stacked"></i>  INVENTORY  </a></li> 
+          <li><a href="{{url_for('addprd')}}"> <i class="fa-solid fa-square-plus"></i> ADD PRODUCT</a></li>
+          <li><a href="{{url_for('transaction')}}"><i class="fa-solid fa-clipboard-list"></i>  TRANSACTIONS </a></li>
+          <li><a href="{{url_for('aboutus')}}"> <i class="fa-regular fa-address-card"></i> ABOUT US  </a></li> 
+          <li><a href="{{url_for('logout')}}"> <i class="fa-solid fa-right-from-bracket"></i> LOGOUT  </a></li> 
+        </ul> 
+      </div> 
+      <div class="main_content"> 
+        <div class="header">DASHBOARD</div> 
+        <div class="info"> 
+
+
+          <script type="text/javascript" src="https://www.gstatic.com/charts/loader.js"></script>
+          <script type="text/javascript">
+            google.charts.load('current', {'packages':['corechart']});
+            google.charts.setOnLoadCallback(drawChart);
+          
+            function drawChart() {
+              var data = google.visualization.arrayToDataTable([
+                ['Product', 'Quantity'],
+                ['Drinks', {{drinksCount[0]}}],
+                ['Snacks', {{snackCount[0]}}],
+                ['Daily Products', {{dairyProductcount[0]}}],
+                ['Condiments and Sauces', {{condimentSaucesCount[0]}}],
+                ['Household Items', {{householdItemCount[0]}}],
+                ['Personal Care Products', {{personalCareCount[0]}}],
+              ]);
+          
+              var options = {
+                title: 'PRODUCT OVERALL STOCKS',
+                backgroundColor: '#dfeae9',
+                pieSliceText: 'label',
+                legend: {
+                  position: 'bottom',
+                  alignment: 'center',
+                  textStyle: {
+                    color: 'black',
+                  }
+                }
+              };
+          
+              var chart = new google.visualization.PieChart(document.getElementById('piechart'));
+          
+              // Identify products with low stock and display warnings
+              var warningList = document.getElementById('warnings');
+              warningList.innerHTML = ''; // Clear any existing warnings
+          
+              for (var i = 0; i < data.getNumberOfRows(); i++) {
+                if (data.getValue(i, 1) < 50) { // Check if quantity is less than 50
+                  var productName = data.getValue(i, 0);
+                  var warningItem = document.createElement('li');
+                  warningItem.style.color = 'red';
+                  warningItem.innerHTML = `⚠️ Low stock: ${productName}`;
+                  warningList.appendChild(warningItem); // Add warning to the list
+                }
+              }
+          
+              chart.draw(data, options);
+            }
+          </script>
+          
+          <div id="piechart" style="width: 900px; height: 700px;"></div>
+          <ul id="warnings"></ul>
+          
+
+ 
+    <script> 
+        document.querySelector('.hamburger-menu').addEventListener('click', function () { 
+          document.querySelector('.sidebar').classList.toggle('active'); 
+        }); 
+      </script> 
+
+        </div>
+    </div>
+
+
+
+  </body> 
+</html>
