+++ templates/delete.html
@@ -1,0 +1,27 @@
+<!DOCTYPE html>
+<html lang="en">
+  <head>
+    <meta charset="utf-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0, maximum-scale=1.0, user-scalable=no">
+    <title>Confirm Deletion</title>
+    <link rel="stylesheet" href="static/delete.css" >
+  </head>
+  <body>
+    <h2>Confirm Deletion</h2>
+    <p>Are you sure you want to delete product {{ productID }}?</p>
+    {% with messages = get_flashed_messages() %}
+      {% if messages %}
+        <ul>
+          {% for message in messages %}
+            <li>{{ message }}</li>
+          {% endfor %}
+        </ul>
+      {% endif %}
+    {% endwith %}
+    <form action="/delete_product" method="POST">
+      <input type="hidden" name="productID" value="{{ productID }}">
+      <button class="button" type="submit" name="confirm" value="yes">Yes, delete</button>
+      <button type="submit" name="confirm" value="no">No, cancel</button>
+    </form>
+  </body>
+</html>
