+++ templates/inventory.html
@@ -1,0 +1,115 @@
+<!DOCTYPE html> 
+<html lang="en"> 
+  <head> 
+    <meta charset="utf-8"> 
+    <meta name="viewport" content="width=device-width, initial-scale=1.0, maximum-scale=1.0, user-scalable=no"> 
+    <title>INVENTORY</title> 
+    <link rel="stylesheet" href="static/inv.css" > 
+ 
+    <link rel="preconnect" href="https://fonts.googleapis.com"> 
+    <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin> 
+    <link href="https://fonts.googleapis.com/css2?family=Kanit:ital,wght@0,100;0,200;0,300;0,400;0,500;0,600;0,700;0,800;0,900;1,100;1,200;1,300;1,400;1,500;1,600;1,700;1,800;1,900&display=swap" rel="stylesheet"> 
+    <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.1.1/css/all.min.css" /> 
+     
+ 
+ 
+ 
+  </head> 
+  <body> 
+    <div class="hamburger-menu"> 
+        <i class="fas fa-bars"></i> 
+    </div> 
+    <div class="wrapper"> 
+      <div class="sidebar"> 
+        <h3>INVENTORY MANAGEMENT</h3> 
+        <ul> 
+          <li><a href="{{url_for('dashboard')}}"> <i class="fa-solid fa-chart-pie"></i> DASHBOARD</a></li>
+          <li><a href="{{url_for('inventory')}}"> <i class="fa-solid fa-boxes-stacked"></i>  INVENTORY  </a></li> 
+          <li><a href="{{url_for('addprd')}}"> <i class="fa-solid fa-square-plus"></i> ADD PRODUCT</a></li>
+          <li><a href="{{url_for('transaction')}}"><i class="fa-solid fa-clipboard-list"></i>  TRANSACTIONS </a></li>
+          <li><a href="{{url_for('aboutus')}}"> <i class="fa-regular fa-address-card"></i> ABOUT US  </a></li> 
+          <li><a href="{{url_for('logout')}}"> <i class="fa-solid fa-right-from-bracket"></i> LOGOUT  </a></li> 
+        </ul> 
+      </div> 
+      <div class="main_content"> 
+        <div class="header">INVENTORY</div> 
+        <div class="info"> 
+ 
+ 
+ 
+          {% with messages = get_flashed_messages() %} 
+            {% if messages %} 
+              <ul> 
+                {% for message in messages %} 
+                  <li>{{ message }}</li> 
+                {% endfor %} 
+              </ul> 
+            {% endif %} 
+          {% endwith %} 
+ 
+ 
+ 
+ 
+ 
+ 
+ 
+ 
+ 
+           
+          <div class="search-bar"> 
+            <form action="/search" method="POST"> 
+              <input type="text" name="search_query" placeholder="Search products..."> 
+              <button type="submit">Search</button> 
+            </form> 
+          </div> 
+          {% for category, items in products.items() %} 
+            <h2>{{ category }}</h2> 
+            <table> 
+              <th>Product ID</th> 
+              <th>Product Name</th> 
+              <th>Product Quantity</th> 
+              <th>Product Price</th> 
+              <th>Operation</th> 
+              {% for item in items %} 
+                <tr> 
+                  <form action="/update" method="POST"> 
+                    <td> 
+                      <label>{{ item[0] }}</label> 
+                      <input type="hidden" name="productID" value="{{ item[0] }}"> 
+                    </td> 
+                    <td> 
+                      <label>{{ item[1] }}</label> 
+                      <input type="hidden" name="product_name" value="{{ item[1] }}"> 
+                    </td> 
+                    <td> 
+                      <label>{{ item[2] }}</label> 
+                      <input type="hidden" name="Qty" value="{{ item[2] }}"> 
+                    </td> 
+                    <td> 
+                      <label><i class="fa-solid fa-peso-sign"></i>{{ item[3] }}</label> 
+                      <input type="hidden" name="price" value="{{ item[3] }}"> 
+                    </td> 
+                    <td> 
+                      <button type="submit" name="action" value="updater"> update product </button> 
+                      <button type="submit" name="action" value="deleter"> remove product </button> 
+                    </td> 
+                  </form> 
+                </tr> 
+              {% endfor %} 
+            </table> 
+          {% endfor %} 
+        </div> 
+      </div> 
+    </div> 
+ 
+    <script> 
+        document.querySelector('.hamburger-menu').addEventListener('click', function () { 
+          document.querySelector('.sidebar').classList.toggle('active'); 
+        }); 
+      </script> 
+
+
+
+
+  </body> 
+</html>
