+++ templates/signup.html
@@ -1,0 +1,27 @@
+<!DOCTYPE html>
+<html lang="en">
+  <head>
+    <meta charset="utf-8">
+    <title>SIGN UP</title>
+    <link rel="stylesheet" href="static/login.css" >
+    <link rel="preconnect" href="https://fonts.googleapis.com">
+    <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
+    <link href="https://fonts.googleapis.com/css2?family=Oswald:wght@200..700&display=swap" rel="stylesheet">
+  </head>
+  <body>
+    <div class="header">
+      <h1>WELCOME TO INVENTORY MANAGEMENT SYSTEM.</h1>
+    </div>
+
+    <form action="/signup_process" method="post">
+      <h3>CREATE AN ACCOUNT.</h3>
+      <input type="text" placeholder="username" name="username" required>
+      <br>
+      <input type="password" placeholder="password" name="password" required>
+      <br>
+      <input type="password" placeholder="confirm password" name="confirm_password" required>
+      <br>
+      <button> Sign up</button>
+    </form>
+  </body>
+</html>
